# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sway cruise booking and operations API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Cruise {
  id: number;
  name: string;
  tagline: string;
  route: string;
  durationHours: number;
  capacity: number;
  pricePerPerson: number;
  rating: number;
  imageUrl: string;
  departure: string;
  highlights: string[];
}

export interface BookingInput {
  cruiseId: number;
  /** @minLength 2 */
  guestName: string;
  guestEmail: string;
  guestPhone?: string;
  departureDate: string;
  /**
     * @minimum 1
     * @maximum 12
     */
  passengers: number;
  notes?: string;
}

export interface Booking {
  id: number;
  reference: string;
  cruiseId: number;
  cruiseName: string;
  guestName: string;
  guestEmail: string;
  /** @nullable */
  guestPhone?: string | null;
  departureDate: string;
  passengers: number;
  total: number;
  status: string;
  paymentStatus: string;
  createdAt: string;
}

export type AdminBooking = Booking & ({
  /** @nullable */
  vesselId?: number | null;
  /** @nullable */
  vesselName?: string | null;
  paymentMethod: string;
  departureTime: string;
});

export type BookingUpdateStatus = typeof BookingUpdateStatus[keyof typeof BookingUpdateStatus];


export const BookingUpdateStatus = {
  requested: 'requested',
  confirmed: 'confirmed',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type BookingUpdatePaymentStatus = typeof BookingUpdatePaymentStatus[keyof typeof BookingUpdatePaymentStatus];


export const BookingUpdatePaymentStatus = {
  unpaid: 'unpaid',
  paid: 'paid',
  overdue: 'overdue',
  refunded: 'refunded',
} as const;

export interface BookingUpdate {
  status?: BookingUpdateStatus;
  /** @nullable */
  vesselId?: number | null;
  departureDate?: string;
  paymentStatus?: BookingUpdatePaymentStatus;
}

export type VesselStatus = typeof VesselStatus[keyof typeof VesselStatus];


export const VesselStatus = {
  available: 'available',
  on_cruise: 'on_cruise',
  maintenance: 'maintenance',
  attention: 'attention',
} as const;

export interface Vessel {
  id: number;
  name: string;
  model: string;
  registration: string;
  capacity: number;
  status: VesselStatus;
  location: string;
  mileage: number;
  lastService: string;
  nextService: string;
  fuelPercent: number;
  imageUrl: string;
}

export type VesselUpdateStatus = typeof VesselUpdateStatus[keyof typeof VesselUpdateStatus];


export const VesselUpdateStatus = {
  available: 'available',
  on_cruise: 'on_cruise',
  maintenance: 'maintenance',
  attention: 'attention',
} as const;

export interface VesselUpdate {
  status?: VesselUpdateStatus;
  nextService?: string;
}

export interface TrendPoint {
  label: string;
  value: number;
}

export type OperationalAlertSeverity = typeof OperationalAlertSeverity[keyof typeof OperationalAlertSeverity];


export const OperationalAlertSeverity = {
  critical: 'critical',
  warning: 'warning',
  info: 'info',
} as const;

export interface OperationalAlert {
  id: number;
  severity: OperationalAlertSeverity;
  category: string;
  title: string;
  detail: string;
  /** @nullable */
  vesselName?: string | null;
  createdAt: string;
  resolved: boolean;
}

export interface AdminOverview {
  totalBookings: number;
  activeVessels: number;
  revenue: number;
  pendingRequests: number;
  occupancyPercent: number;
  revenueTrend: TrendPoint[];
  bookingTrend: TrendPoint[];
  alerts: OperationalAlert[];
}

export interface PaymentBreakdown {
  method: string;
  count: number;
  amount: number;
}

export interface RevenueReport {
  grossRevenue: number;
  collectedRevenue: number;
  outstanding: number;
  monthly: TrendPoint[];
  paymentMethods: PaymentBreakdown[];
}

export type ServiceHealthStatus = typeof ServiceHealthStatus[keyof typeof ServiceHealthStatus];


export const ServiceHealthStatus = {
  operational: 'operational',
  degraded: 'degraded',
  outage: 'outage',
} as const;

export interface ServiceHealth {
  name: string;
  status: ServiceHealthStatus;
  latencyMs: number;
  checkedAt: string;
}

export interface MonitoringReport {
  generatedAt: string;
  fleetAvailability: number;
  openAlerts: number;
  overdueReturns: number;
  maintenanceDue: number;
  paymentIssues: number;
  services: ServiceHealth[];
  alerts: OperationalAlert[];
}

export interface ActivityEvent {
  id: number;
  kind: string;
  title: string;
  detail: string;
  createdAt: string;
}

export type ListAdminBookingsParams = {
search?: string;
status?: ListAdminBookingsStatus;
};

export type ListAdminBookingsStatus = typeof ListAdminBookingsStatus[keyof typeof ListAdminBookingsStatus];


export const ListAdminBookingsStatus = {
  requested: 'requested',
  confirmed: 'confirmed',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;