# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sway cruise booking and operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/fleet/{vesselId}:
    patch:
      operationId: updateVessel
      tags: [operations]
      summary: Update vessel availability or service schedule
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VesselUpdate"
      responses:
        "200":
          description: Updated vessel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vessel"
        "404":
          description: Vessel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/bookings:
    get:
      operationId: listAdminBookings
      tags: [operations]
      summary: List and filter booking requests
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [requested, confirmed, in_progress, completed, cancelled]
      responses:
        "200":
          description: Booking records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminBooking"
  /admin/bookings/{bookingId}:
    patch:
      operationId: updateBooking
      tags: [operations]
      summary: Approve, cancel, extend, or assign a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBooking"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Booking:
      type: object
      properties:
        id: { type: integer }
        reference: { type: string }
        cruiseId: { type: integer }
        cruiseName: { type: string }
        guestName: { type: string }
        guestEmail: { type: string }
        guestPhone: { type: ["string", "null"] }
        departureDate: { type: string, format: date }
        passengers: { type: integer }
        total: { type: number }
        status: { type: string }
        paymentStatus: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, reference, cruiseId, cruiseName, guestName, guestEmail, departureDate, passengers, total, status, paymentStatus, createdAt]
    AdminBooking:
      allOf:
        - $ref: "#/components/schemas/Booking"
        - type: object
          properties:
            vesselId: { type: ["integer", "null"] }
            vesselName: { type: ["string", "null"] }
            paymentMethod: { type: string }
            departureTime: { type: string }
          required: [paymentMethod, departureTime]
    BookingUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [requested, confirmed, in_progress, completed, cancelled]
        vesselId:
          type: ["integer", "null"]
        departureDate:
          type: string
          format: date
        paymentStatus:
          type: string
          enum: [unpaid, paid, overdue, refunded]
    Vessel:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        model: { type: string }
        registration: { type: string }
        capacity: { type: integer }
        status:
          type: string
          enum: [available, on_cruise, maintenance, attention]
        location: { type: string }
        mileage: { type: integer }
        lastService: { type: string, format: date }
        nextService: { type: string, format: date }
        fuelPercent: { type: integer }
        imageUrl: { type: string }
      required: [id, name, model, registration, capacity, status, location, mileage, lastService, nextService, fuelPercent, imageUrl]
    VesselUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [available, on_cruise, maintenance, attention]
        nextService:
          type: string
          format: date
...