# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sway cruise booking and operations API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  reference: string;
  cruiseId: number;
  cruiseName: string;
  guestName: string;
  guestEmail: string;
  /** @nullable */
  guestPhone?: string | null;
  departureDate: Date;
  passengers: number;
  total: number;
  status: string;
  paymentStatus: string;
  createdAt: Date;
}