# lib/api-zod/src/generated/types/bookingUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sway cruise booking and operations API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingUpdatePaymentStatus } from './bookingUpdatePaymentStatus';
import type { BookingUpdateStatus } from './bookingUpdateStatus';

export interface BookingUpdate {
  status?: BookingUpdateStatus;
  /** @nullable */
  vesselId?: number | null;
  departureDate?: Date;
  paymentStatus?: BookingUpdatePaymentStatus;
}