# lib/api-zod/src/generated/types/vessel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Sway cruise booking and operations API
 * OpenAPI spec version: 0.1.0
 */
import type { VesselStatus } from './vesselStatus';

export interface Vessel {
  id: number;
  name: string;
  model: string;
  registration: string;
  capacity: number;
  status: VesselStatus;
  location: string;
  mileage: number;
  lastService: Date;
  nextService: Date;
  fuelPercent: number;
  imageUrl: string;
}